requirements.txt
opencv-python-headless==4.9.0.80
colorthief==0.2.1
python-dotenv==1.0.1
anthropic==0.25.2
numpy==1.26.4
easyocr==1.7.1
torch==2.2.2
gradio==4.26.0
lightning==2.2.5
